Using SDE hosted service instead of AGOL service
- Using common application definitions for both development and production.
- Updated recycling metadata schema for SDE hosted service
- Filtering map recycling locations by public view only
