+++ apps/ues-recycling-angular/src/environments/environment.prod.ts
@@ -2,208 +2,6 @@ export const environment = {
   production: true
 };
 
-import { Connections, Definitions as d } from './definitions';
-import { SearchSource, SearchSourceQueryParamsProperties } from '@tamu-gisc/search';
-import { LayerSource, LegendItem } from '@tamu-gisc/common/types';
-import { Popups } from '@tamu-gisc/aggiemap';
-
-import esri = __esri;
-
 export * from './definitions';
 
 export const apiUrl = 'https://nodes.geoservices.tamu.edu/api/ues/recycling/';
-
-const commonLayerProps = {
-  outFields: ['*'],
-  minScale: 100000,
-  maxScale: 0,
-  elevationInfo: { mode: 'relative-to-ground', offset: 1 } as esri.FeatureLayerElevationInfo,
-  popupEnabled: false
-};
-
-// Persistent layer definitions that will be processed by a factory and added to the map.
-export const LayerSources: LayerSource[] = [
-  {
-    type: 'feature',
-    id: d.BUILDINGS.layerId,
-    title: d.BUILDINGS.name,
-    url: d.BUILDINGS.url,
-    popupComponent: Popups.BuildingPopupComponent,
-    listMode: 'hide',
-    loadOnInit: true,
-    visible: true,
-    layerIndex: 2,
-    native: {
-      ...commonLayerProps,
-      renderer: {
-        type: 'simple',
-        symbol: {
-          type: 'simple-fill',
-          style: 'solid',
-          color: [0, 0, 0, 0.01],
-          outline: {
-            width: '0'
-          }
-        }
-      }
-    }
-  },
-  {
-    type: 'feature',
-    id: d.RECYCLING.layerId,
-    title: d.RECYCLING.name,
-    url: d.RECYCLING.url,
-    popupComponent: Popups.BuildingPopupComponent,
-    listMode: 'hide',
-    loadOnInit: true,
-    visible: true,
-    layerIndex: 2,
-    native: {
-      ...commonLayerProps
-    }
-  },
-  {
-    type: 'graphic',
-    id: 'selection-layer',
-    title: 'Selected Buildings',
-    category: 'Infrastructure',
-    listMode: 'hide',
-    loadOnInit: false,
-    visible: true,
-    popupComponent: Popups.BasePopupComponent
-  }
-];
-
-// Static legend items
-export const LegendSources: LegendItem[] = [
-  {
-    id: 'university-building-legend',
-    title: 'University Building',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAJBJREFUOI1jYaAyYIExAtnZGygxaP3Pnw1wAwPZ2RuE7azruYVFyDLsw/2HDCkvXhirKCvPgruQW1iEQUJJmWwXMv386cPAzOzDQlgpaWDUwFEDRw0cZgZ+uP+QbEO+vn3DwIts4PqfPxsCL1xg+HThAgMDAwODsISEMamGimtooLoQVuIyMDAwMDwkw7VQPQDNmyRazmV6EgAAAABJRU5ErkJggg=='
-  },
-  {
-    id: 'non-university-building-legend',
-    title: 'Non-University Building',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHFJREFUOI1jYaAyYIExmpqaGigxqK6urgFuYFNTU4OCBGe9EC8bWYY9ff2doampiaGurq4B7kIhXjYGZWk+sl348sMvBrgLqQlGDRw1cNTAYWbg09ffyTbk3edfqAbW1dU1NDU1wYsgcgBKAYssQCkAABa3ILBI1xApAAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'off-campus-building-legend',
-    title: 'Off Campus Building',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGdJREFUOI3t1KENADEMA8CADhBc2H0802cnLxVUdYRHH7U0LXrVKAo4GbnI4ZTvIGk7EAALkKTVWh9VTWHuLiQFgEVDVZXWWrph712i4clc8IIX/Bno7mlkjLGCAIxkTFAmy8DOj9281XsgH/GbKEkAAAAASUVORK5CYII='
-  },
-  {
-    id: 'garage-parking-legend',
-    title: 'Garage Parking',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGtJREFUOI3t1LEOgCAMBNAOTMBgogNfezb3v4YQxdnJRldkMtzUdHi56Zx0jrsPkvoFAqAGktRlntYYfBO25SIkBYBawxi8pLQ0N9z3KtawZwY4wAH+DNxyaUaOer5BAErSJqglr4F9Pr7mAt2qILaQAOa8AAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'restricted-legend',
-    title: 'Restricted 6am to 6pm',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAABHNCSVQICAgIfAhkiAAAAFtJREFUOI1jYKAyYIQxHB0dGygxaP/+/Q0MDAwMLDDDQkMC6+OfviDLsFmiwnBDmWCC5BrGwMDAkPb6LZzNhEcdWWDUwFEDRw0cZgbCiiBywEJpCTib6gUs1QEATW0WjAt6JIYAAAAASUVORK5CYII='
-  },
-  {
-    id: 'paved-surface-legend',
-    title: 'Paved Surface',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGVJREFUOI3t1CEOACEMBMAKdF+A4aN7zf62AsMLTl0DtqAurGoqJqu2yOGU7yBpOxAAC5Ck1VofVU1h7i4kBYBFQ1WV1lq6Ye9douHJXPCCF/wZ6O5pZIyxggCMZExQJsvAzo/dvKA4IK9bweSrAAAAAElFTkSuQmCC'
-  },
-  {
-    id: 'sidewalk-legend',
-    title: 'Sidewalk',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAEtJREFUOI1jYaAyYIExPr641UCJQfwSag1wA6GG1VNi4McXtxj4JdQaWAgrJQ2MGjhq4KiBowbS0UB+CbWGjy9uUWQQSgGLLEApAAAvoxCdulpoFwAAAABJRU5ErkJggg=='
-  },
-  {
-    id: 'railroad-legend',
-    title: 'Railroad',
-    src:
-      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHhJREFUOI3tkrENgDAMBL+I9EUmYA7WYCfCTLBG5qBz/x0VRUAkAtIg8q3t+7dlh8pyDfgnIMnwBiQpJEBJI8mpMDQAWEnGYkIA0Xs/m9llM8kkTQlYRQnQzEJuHUk9gG5PeqiFBEhyKbnnzE4Jc3e5ow89dgM+1gZwZSYCkyusZQAAAABJRU5ErkJggg=='
-  }
-];
-
-const commonQueryParams: Partial<SearchSourceQueryParamsProperties> = {
-  f: 'json',
-  resultRecordCount: 5,
-  outFields: '*',
-  outSR: 4326,
-  returnGeometry: true,
-  spatialRel: 'esriSpatialRelIntersects'
-};
-
-// Search sources used for querying features.
-export const SearchSources: SearchSource[] = [
-  {
-    source: 'building',
-    name: 'Building',
-    url: `${Connections.basemapUrl}/1`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['Number', 'BldgAbbr', 'BldgName'],
-        operators: ['LIKE', 'LIKE', 'LIKE'],
-        wildcards: ['includes', 'includes', 'includes'],
-        transformations: ['UPPER', 'UPPER', 'UPPER']
-      },
-      scoringWhere: {
-        keys: ['BldgName'],
-        operators: ['LIKE'],
-        wildcards: ['startsWith'],
-        transformations: ['UPPER']
-      }
-    },
-    scoringKeys: ['attributes.BldgAbbr', 'attributes.Number', 'attributes.BldgName'],
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
-    popupComponent: Popups.BuildingPopupComponent,
-    searchActive: true
-  },
-  {
-    source: 'building-exact',
-    name: 'Building',
-    url: `${Connections.basemapUrl}/1`,
-    queryParams: {
-      ...commonQueryParams,
-      where: {
-        keys: ['Number'],
-        operators: ['=']
-      }
-    },
-    featuresLocation: 'features',
-    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
-    popupComponent: Popups.BuildingPopupComponent,
-    searchActive: false
-  }
-];
-
-export const SelectionSymbols = {
-  polygon: {
-    type: 'simple-fill',
-    style: 'solid',
-    color: [252, 227, 0, 0.55],
-    outline: {
-      color: [252, 227, 0, 0.8],
-      width: '2px'
-    }
-  },
-  point: {
-    type: 'simple-marker',
-    style: 'circle',
-    size: 8,
-    outline: {
-      width: 2
-    }
-  },
-  multipoint: {
-    type: 'simple-marker',
-    style: 'circle',
-    size: 8,
-    outline: {
-      width: 2
-    }
-  }
-};
